+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Added `tmp_path` helper function
+- Added parsing command options with values enclosed in double quotes
+- Added type `array` for Console options, allow to set multiple values
 - Added `formGroup` template to `FormHelper`
 - Added `formCheck` template to `FormHelper`